REST_API/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tk==0.1.0
urllib3==1.26.12
virtualenv==20.16.4
whitenoise==6.2.0
